kafka-orders/requirements.txt
kafka-python==2.1.0
fastavro==1.9.3
